+++ src/interfaces/user.interfaces.ts
@@ -1,0 +1,21 @@
+export interface User {
+  id: number;
+  name: string;
+  surname: string;
+  age: number;
+}
+
+export interface CreateUserResult {
+  result: User[];
+  message: string;
+}
+
+export interface UpdateUserResult {
+  result: User[];
+  message: string;
+}
+
+export interface DeleteUserResult {
+  result: User[];
+  message: string;
+}
